Use typedef for WsrepLocalState everywhere. [#135653767]

+++ src/github.com/cloudfoundry-incubator/galera-healthcheck/config/config_test.go
@@ -163,7 +163,7 @@ var _ = Describe("Config", func() {
 			}
 
 			state := domain.DBState{
-				WsrepLocalState: uint(ls),
+				WsrepLocalState: ls,
 				ReadOnly:        readOnly,
 			}
